fix(symthaea): resolve all 39 clippy warnings — zero warnings now
Privacy fixes (14): make types pub that were exposed through public APIs:
- CycleUrgency, ActionHint, PsiAttestationRecord, MoralJudgmentSummary
- ActiveInferenceBridgeStats, MemoryBridgeStats, FlowTemporalSummary
- ReflectionSummary, ReflectionThresholds, Recommendation

Dead code fixes (20): #[allow(dead_code)] on dormant planned items:
- CognitiveLoopExecutor, CodeTaskDetector, CausalGoalAnalyzer
- CycleMetadataCompact, from_arousal, phi_modulation, etc.

Auto-fix (5): remove unnecessary as-usize casts in cycle_subsystems.rs

+++ symthaea/src/cognitive_loop/cycle_subsystems.rs
@@ -170,7 +170,7 @@ impl CognitiveLoopService {
         let _t = Instant::now();
         let (holographic_unity, holographic_binding) =
             if let Some(ref mut ha) = self.holographic_analyzer {
-                if self.stats.total_cycles % (19 * budget_interval_mult as usize) == 0 {
+                if self.stats.total_cycles % (19 * budget_interval_mult) == 0 {
                     let content: Vec<f64> = (0..64)
                         .map(|i| {
                             if hv16_cached.get_bit(i) != 0 {
@@ -216,7 +216,7 @@ impl CognitiveLoopService {
         let _t = Instant::now();
         let (consciousness_gradient_magnitude, consciousness_limiting_component) =
             if let Some(ref dc) = self.differentiable_consciousness {
-                if self.stats.total_cycles % (23 * budget_interval_mult as usize) == 0
+                if self.stats.total_cycles % (23 * budget_interval_mult) == 0
                     && self.stats.total_cycles > 0
                 {
                     use crate::consciousness::consciousness_equation_v2::{
@@ -274,7 +274,7 @@ impl CognitiveLoopService {
         let (affect_cons_valence, affect_cons_arousal) =
             if let Some(ref mut ac) = self.affective_consciousness {
                 ac.decay(0.05);
-                if self.stats.total_cycles % (11 * budget_interval_mult as usize) == 0 {
+                if self.stats.total_cycles % (11 * budget_interval_mult) == 0 {
                     let valence = 1.0 - 2.0 * prediction_error;
                     let base_affect = crate::consciousness::affective_consciousness::CoreAffect {
                         valence,
